Fix bug with search dropdown
When navigating away from ExploreCurriculum using Turbolinks,
and then going back to E.C., the search dropdown wouldn't work
anymore.

This happened because the search dropdown component was returning
null when the dropdown was hidden.

React renders nulls with a <noscript> tag, but Turbolinks 2.5.3
removes <noscript> tags. Every time this happened, React noticed
the HTML was mutated and freaked out.