Fix bug in EC page
For some reason Rails isn't always picking up the "Accept"
header. Sometimes it works, sometimes it doesn't.

Could be something in EBS, Varnish or Nginx, or some bug
in the original JS request I have missed.

In any case, the usage for the js-routes helper is wrong,
and by fixing it to use the proper format it will always
work (format will be determined by URL extension rather
than the "Accept" header).