Fix problem in CG downloads
Because we're using the staging environment as literally a "staging"
area before moving data changes into production (something that should
certainly be revised), some weird bugs show up when data depends
on external services.

The content guide filename column in the database is used to locate
the file in S3 but also as an indicator of whether this file should
be refreshed. When a content guide is reimported, and its PDF rendering
should be regenerated, we clear out this column so the next download
will refresh the PDF. Clearing out the column deletes the file on S3.

These PDFs are stored in S3 for both staging and production.
When a content guide is reimported in staging, it clears out the column
on staging and deletes the file from S3. However, the database in
production remains untouched. Production doesn't know the PDF was
deleted and should be refreshed, so it points the user at a file
that doesn't exist (S3 error).

I've modified the CG code to do a few things:

1) Don't delete previous versions of PDFs.
2) Version out each PDF download so none will be overwritten.
3) When downloading a PDF, check whether the CG content version matches
   the PDF version. If it doesn't, generate the PDF.